==> 4 SCENIC analysis.py <==
# ExN_CA
pyscenic grn --num_workers 30 --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/adj.ExN_CA.tsv --method grnboost2 /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/matrix_ExN_CA.csv /data/whu/home/whr_brain_project/data/scenic/resources/allTFs_hg38.txt
pyscenic ctx /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/adj.ExN_CA.tsv /data/whu/home/whr_brain_project/data/scenic/databases/hg38_500bp_up_100bp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather /data/whu/home/whr_brain_project/data/scenic/databases/hg38_10kbp_up_10kbp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather --annotations_fname /data/whu/home/whr_brain_project/data/scenic/resources/motifs-v10nr_clust-nr.hgnc-m0.001-o0.0.tbl --expression_mtx_fname /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/matrix_ExN_CA.csv --mode dask_multiprocessing --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/reg.ExN_CA.csv --num_workers 30 --mask_dropouts
pyscenic aucell /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/matrix_ExN_CA.csv /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/reg.ExN_CA.csv --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_CA/auc_mtx.ExN_CA.csv --num_workers 20

# ExN_DG
pyscenic grn --num_workers 30 --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/adj.ExN_DG.tsv --method grnboost2 /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/matrix_ExN_DG.csv /data/whu/home/whr_brain_project/data/scenic/resources/allTFs_hg38.txt
pyscenic ctx /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/adj.ExN_DG.tsv /data/whu/home/whr_brain_project/data/scenic/databases/hg38_500bp_up_100bp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather /data/whu/home/whr_brain_project/data/scenic/databases/hg38_10kbp_up_10kbp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather --annotations_fname /data/whu/home/whr_brain_project/data/scenic/resources/motifs-v10nr_clust-nr.hgnc-m0.001-o0.0.tbl --expression_mtx_fname /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/matrix_ExN_DG.csv --mode dask_multiprocessing --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/reg.ExN_DG.csv --num_workers 30 --mask_dropouts
pyscenic aucell /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/matrix_ExN_DG.csv /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/reg.ExN_DG.csv --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ExN_DG/auc_mtx.ExN_DG.csv --num_workers 20

# InN
pyscenic grn --num_workers 30 --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/adj.InN.tsv --method grnboost2 /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/matrix_InN.csv /data/whu/home/whr_brain_project/data/scenic/resources/allTFs_hg38.txt
pyscenic ctx /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/adj.InN.tsv /data/whu/home/whr_brain_project/data/scenic/databases/hg38_500bp_up_100bp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather /data/whu/home/whr_brain_project/data/scenic/databases/hg38_10kbp_up_10kbp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather --annotations_fname /data/whu/home/whr_brain_project/data/scenic/resources/motifs-v10nr_clust-nr.hgnc-m0.001-o0.0.tbl --expression_mtx_fname /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/matrix_InN.csv --mode dask_multiprocessing --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/reg.InN.csv --num_workers 30 --mask_dropouts
pyscenic aucell /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/matrix_InN.csv /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/reg.InN.csv --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_InN/auc_mtx.InN.csv --num_workers 20

# ASC
pyscenic grn --num_workers 30 --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/adj.ASC.tsv --method grnboost2 /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/matrix_ASC.csv /data/whu/home/whr_brain_project/data/scenic/resources/allTFs_hg38.txt
pyscenic ctx /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/adj.ASC.tsv /data/whu/home/whr_brain_project/data/scenic/databases/hg38_500bp_up_100bp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather /data/whu/home/whr_brain_project/data/scenic/databases/hg38_10kbp_up_10kbp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather --annotations_fname /data/whu/home/whr_brain_project/data/scenic/resources/motifs-v10nr_clust-nr.hgnc-m0.001-o0.0.tbl --expression_mtx_fname /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/matrix_ASC.csv --mode dask_multiprocessing --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/reg.ASC.csv --num_workers 30 --mask_dropouts
pyscenic aucell /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/matrix_ASC.csv /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/reg.ASC.csv --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ASC/auc_mtx.ASC.csv --num_workers 20

# MG
pyscenic grn --num_workers 30 --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/adj.MG.tsv --method grnboost2 /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/matrix_MG.csv /data/whu/home/whr_brain_project/data/scenic/resources/allTFs_hg38.txt
pyscenic ctx /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/adj.MG.tsv /data/whu/home/whr_brain_project/data/scenic/databases/hg38_500bp_up_100bp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather /data/whu/home/whr_brain_project/data/scenic/databases/hg38_10kbp_up_10kbp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather --annotations_fname /data/whu/home/whr_brain_project/data/scenic/resources/motifs-v10nr_clust-nr.hgnc-m0.001-o0.0.tbl --expression_mtx_fname /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/matrix_MG.csv --mode dask_multiprocessing --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/reg.MG.csv --num_workers 30 --mask_dropouts
pyscenic aucell /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/matrix_MG.csv /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/reg.MG.csv --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_MG/auc_mtx.MG.csv --num_workers 20

# ODC
pyscenic grn --num_workers 30 --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/adj.ODC.tsv --method grnboost2 /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/matrix_ODC.csv /data/whu/home/whr_brain_project/data/scenic/resources/allTFs_hg38.txt
pyscenic ctx /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/adj.ODC.tsv /data/whu/home/whr_brain_project/data/scenic/databases/hg38_500bp_up_100bp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather /data/whu/home/whr_brain_project/data/scenic/databases/hg38_10kbp_up_10kbp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather --annotations_fname /data/whu/home/whr_brain_project/data/scenic/resources/motifs-v10nr_clust-nr.hgnc-m0.001-o0.0.tbl --expression_mtx_fname /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/matrix_ODC.csv --mode dask_multiprocessing --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/reg.ODC.csv --num_workers 30 --mask_dropouts
pyscenic aucell /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/matrix_ODC.csv /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/reg.ODC.csv --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_ODC/auc_mtx.ODC.csv --num_workers 20

# OPC
pyscenic grn --num_workers 30 --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/adj.OPC.tsv --method grnboost2 /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/matrix_OPC.csv /data/whu/home/whr_brain_project/data/scenic/resources/allTFs_hg38.txt
pyscenic ctx /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/adj.OPC.tsv /data/whu/home/whr_brain_project/data/scenic/databases/hg38_500bp_up_100bp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather /data/whu/home/whr_brain_project/data/scenic/databases/hg38_10kbp_up_10kbp_down_full_tx_v10_clust.genes_vs_motifs.rankings.feather --annotations_fname /data/whu/home/whr_brain_project/data/scenic/resources/motifs-v10nr_clust-nr.hgnc-m0.001-o0.0.tbl --expression_mtx_fname /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/matrix_OPC.csv --mode dask_multiprocessing --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/reg.OPC.csv --num_workers 30 --mask_dropouts
pyscenic aucell /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/matrix_OPC.csv /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/reg.OPC.csv --output /data/whu/home/whr_brain_project_final/results/figure_4_SCENIC/cell_type_matrix/SCENIC_OPC/auc_mtx.OPC.csv --num_workers 20
